python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rake-sdk"
version = "0.1.0"
description = "Python SDK for the rake secure LLM agent sandbox"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["rake", "llm", "agent", "security", "azure", "analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = []

[project.optional-dependencies]
azure = [
    "azure-functions>=1.18.0",
    "azure-storage-blob>=12.19.0",
    "azure-servicebus>=7.11.0",
    "azure-identity>=1.15.0",
    "applicationinsights>=0.11.10",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "httpx>=0.27",
    "PyGithub>=2.2",
]

[tool.hatch.build.targets.wheel]
packages = ["rake_sdk"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
